Challenging Recursion
The purpose of this function is to determine if a list of strings is in alphabetic order, with restrictions on built in functions while retaining efficient and minimal code. The function consumes a list of strings such as (list "Abc" "bc" "charlie" "davis") and ignores the case of the letters. It produces true if the list is in alphabetic order, and false otherwise. The first step was to de-capitalize every string and reconstruct a list that is lower case (4 helper functions required, must convert from list of strings to individual characters to downsize them all), then to compare each individual string with the next in the list.